dev_requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile dev_requirements.in
#
annotated-types==0.7.0
    # via pydantic
astor==0.8.1
    # via flake8-qgis
attrs==23.2.0
    # via flake8-bugbear
black==24.3.0
    # via -r dev_requirements.in
build==1.2.1
    # via pip-tools
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   click-option-group
    #   pip-tools
    #   python-semantic-release
click-option-group==0.5.6
    # via python-semantic-release
coverage[toml]==7.4.4
    # via pytest-cov
dotty-dict==1.3.1
    # via python-semantic-release
flake8==7.0.0
    # via
    #   -r dev_requirements.in
    #   flake8-bugbear
    #   flake8-qgis
flake8-bugbear==24.2.6
    # via -r dev_requirements.in
flake8-qgis==1.0.0
    # via -r dev_requirements.in
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via python-semantic-release
idna==3.7
    # via requests
importlib-resources==6.4.0
    # via python-semantic-release
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via -r dev_requirements.in
jinja2==3.1.4
    # via
    #   jinja2-cli
    #   python-semantic-release
jinja2-cli==0.8.2
    # via -r dev_requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mypy==1.9.0
    # via -r dev_requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==24.0
    # via
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r dev_requirements.in
platformdirs==4.2.0
    # via black
pluggy==1.4.0
    # via pytest
pyclean==3.0.0
    # via -r dev_requirements.in
pycodestyle==2.11.1
    # via flake8
pydantic==2.7.4
    # via python-semantic-release
pydantic-core==2.18.4
    # via pydantic
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via rich
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
pyqt5-stubs==5.15.6.0
    # via
    #   -r dev_requirements.in
    #   qgis-stubs
pytest==8.1.1
    # via
    #   -r dev_requirements.in
    #   pytest-cov
    #   pytest-qgis
pytest-cov==5.0.0
    # via -r dev_requirements.in
pytest-qgis==2.0.0
    # via -r dev_requirements.in
python-gitlab==4.6.0
    # via python-semantic-release
python-semantic-release==9.8.2
    # via -r dev_requirements.in
qgis-stubs==0.2.0.post1
    # via -r dev_requirements.in
requests==2.32.3
    # via
    #   python-gitlab
    #   python-semantic-release
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via python-gitlab
rich==13.7.1
    # via python-semantic-release
shellingham==1.5.4
    # via python-semantic-release
smmap==5.0.1
    # via gitdb
tomlkit==0.12.5
    # via python-semantic-release
typing-extensions==4.11.0
    # via
    #   mypy
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
